pyproject.toml
[build-system]
requires = ["uv_build>=0.9.20,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ktlweb"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "wagtail == 2.16.3",
    "gunicorn",
    "psycopg2-binary == 2.9.11",
    "dj-database-url",
    "boto",
    "arrow",
    "google-api-python-client",
    "markdown",
    "django-storages",
    "django-compressor",
    "pyopenssl",
    "boto3",
]
